feat: move serializers file
